fix for armwrest35b9yel
"armwrest: Subsequent bonus rounds (After rounds 5 and 9) are unplayable because of missing money bag sprite."

the other part of the bug report is out of date, the screenshot doesn't match MAME's current behavior.  The finger still appears cut off, but this is intentional.  I suspect the bag placement is still slightly incorrect and that it's meant to look like the finger is hidden inside the bag material.

I think further improvements would require studying the actual board / schematics to improve the zoom emulation.

git-svn-id: svn://dspnet.fr/mame/trunk@296 749742ba-7341-0410-aadc-df50b521781e