From: JEA Wallace [mailto:jeaw100@york.ac.uk] Subject: AGEMAME upstream patches for u8/0.123
I thought I'd push a few minor coding improvements in now, while I still

have something resembling free time.

Firstly, I've made a few changes to the 6840 interrupt handling, to be 
more like that in the 6821 driver, and to hopefully reduce the IRQ 
swamping that could occur using the old setup. In the process, I've 
added the 6840 IRQ handler to the wire-OR in the MPU4 driver, as that's 
what the schematic shows (there's just one track for each interrupt 
class, all the way down the PCB to the cartridge loading point). Since I

was looking at that, I made MPU4 games that didn't use a CRT as 
screenless as they can get at present, and altered the logging options 
to avoid a redefinition issue if you wanted the core and driver portions

to log at different levels.

Scorpion 2 has also been altered to match the MPU4 coding style, where 
proper handlers re now included for the stepper motor data, which can 
then be expanded on in AGEMAME.



git-svn-id: svn://dspnet.fr/mame/trunk@504 749742ba-7341-0410-aadc-df50b521781e