Implements scanline based timer devices. Converted Centipede, as an example.
To define a scanline timer, use something like this:

MDRV_TIMER_ADD("32V", SCANLINE, generate_interrupt)
MDRV_TIMER_SCANLINE("main", 0, 16)

The first number is the first scanline the timer will fire on, the 2nd number is the increment.
So in this case, the timer will fire on 0, 16, 32, ..., 224, 240, then wrap around
because the screen is defined as 256 lines high.
The current scanline is passed to the callback in its 'param' argument



git-svn-id: svn://dspnet.fr/mame/trunk@987 749742ba-7341-0410-aadc-df50b521781e