From: Cesare Falco [mailto:cesare.falco@gmail.com] Subject: [bugfix] gnga: load correct graphics roms
I used to play the alternate version of Ghost'n'goblins in my arcade,
and I can remember the cross as the last weapon instead of the shield,
which on the other side, is shown in the square below the playing
ground.

In MAME that version was once known as a clone named "gngcross", which
was then renamed to "gnga" starting with 0.35rc2, and merged with
original set.

Unfortunately, rom list for the clone is wrong: it's been pasted from
original set "gng", whilst a couple of graphics roms must be loaded from
"makaimur" instead. You can see this by yourself comparing sources from
0.35rc1 and 0.35rc2.

The attached patch was obtained from SDLMAME on a Linux box, on release
0.123u6. I can't test it compiling MAME code, however this obviously
affects mainline too.

Thanks and keep up your good work!

Cheers,
Cesare Falco
(SDLMAME Ubuntu maintainer).



git-svn-id: svn://dspnet.fr/mame/trunk@1104 749742ba-7341-0410-aadc-df50b521781e