From: Oliver Stoeneberg [mailto:oliverst@online.de] Subject: Machine -> machine cleanups
This patch changes "Machine" to "machine", when the runnin-machine 
parameter is already available. In some cases I could even remove the 
dprecat.h.

I will look for some occurances where it can be easily changed and 
afterward take a step at adding the running_machine parameter in a 
few places.



git-svn-id: svn://dspnet.fr/mame/trunk@1202 749742ba-7341-0410-aadc-df50b521781e