From: JEA Wallace [mailto:jeaw100@york.ac.uk] Subject: Further ACIA changes
Added dynamic control of RX and TX pins to 6850 ACIA emulation - this is
needed for correct operation of the MPU4 Video communications system, 
and may be handy for some MESS systems. To use, set the rx and tx clocks
to 0 in the acia interface, then use acia_tx_clock_in() and 
acia_rx_clock_in() respectively.



git-svn-id: svn://dspnet.fr/mame/trunk@1206 749742ba-7341-0410-aadc-df50b521781e