Removed CPS1 tile kludges and replaced with per-game ROM mapping tables. Ideally, the ROM mapping should be derived from mapping of the PALs found on the B board. However, this is unlikely to happen, so we'll have to live with hardcoded tables.
Note that this is a significan change and some new bugs might be caused by it. Unlikely previous bugs, where extra garbage tiles would appear in the games, the new bugs would be tiles or sprites MISSING--so they will be harder to notice. This would happen because in some cases the 16x16 regions of the gfx ROMs can be used by both scroll2 and sprites, while in other cases they must be separated.

git-svn-id: svn://dspnet.fr/mame/trunk@1249 749742ba-7341-0410-aadc-df50b521781e