Minor rom name update
Changed the Megatouch III roms 36 & 37 to be the same as the new redump based on my own redump of Megatouch IV and the rom names.  IE: megat3.u36 == 9255-01_u36-ro the 01 in the rom name is the "revision" number.  Normally the U36 & U37 doesn't change just U32 & U38 for program revisions.


git-svn-id: svn://dspnet.fr/mame/trunk@1496 749742ba-7341-0410-aadc-df50b521781e