From: Fabio Priuli [mailto:doge.fabio@gmail.com] Subject: a few more
Last batch of diff to replace input_port_read_indexed occurrencies for this u4. I widely tested the changes, so everything should be ok.

As usual 5/6 driver each patch: let me know if something fails to apply.

A final note: I added 2 'empty' ports to bullsdrt in centiped.c. This is a bit of paranoia to cover the possibility that one unknown dip is a flipscreen one (so that centiped_IN2_r has something to read for the second player as well). If bullsdrt is known not to have flip screen code, please remove these.
Anyway tags allowed me to remove an empty 5th input_port in bullsdrt, which was there to have trackballs as 6th input (like centiped & milliped have), so the code is a bit cleaner now :)

Regards
     Fabio

p.s. I also changed a few AM_READ(input_port_N_read) with AM_READ_PORT... as usual more to come after u5 is out!



git-svn-id: svn://dspnet.fr/mame/trunk@1924 749742ba-7341-0410-aadc-df50b521781e