From: Atari Ace [mailto:atari_ace@verizon.net] Subject: [patch] memory_region madness reloaded Hi mamedev,
The memory_region and memory_region_length functions are probably the
two most common functions in MAME that don't take a machine parameter
but should given the syntax of the related apis memory_region_type and
memory_region_flags.  Clearly they didn't get the parameter because of
the sheer number of changes needed to change the apis.  This pair of
patches makes the change, and deals with the consequences.

The first patch makes some changes to help the second patch
along, as well as moves some uses of the memory_region apis within for
loops outside the loops.



git-svn-id: svn://dspnet.fr/mame/trunk@2019 749742ba-7341-0410-aadc-df50b521781e