From: Fabio Priuli [mailto:doge.fabio@gmail.com] Subject: two more patches
Hi,

enclosed please find the following patches

* portM.diff which completes the basic port cleanup in 
drivers starting with M. It's mostly conversion of 8bit 
-> 16/32bit inputs and usage of AM_READ_PORT in memory maps 
(plus a couple of eeprom_bit_r replaced with PORT_CUSTOM 
and daitorid_soundstatus_r in metro.c converted to a 
CUSTOM_INPUT as well)

* avgdvg.diff converts most of the games checking avgdvg_done() 
in inputs to use a CUSTOM_INPUT (which I added to video/avgdvg.c). 
avgdvg_done() is still in use in mhavoc.c (which requires some 
more changes, so I'll take a look later) and in omegrace.c (which 
reads it directly and not through inputs). While at it, I've 
also done a small cleanup of inputs in starwars.c and tempest.c.

Regards,
     Fabio



git-svn-id: svn://dspnet.fr/mame/trunk@2522 749742ba-7341-0410-aadc-df50b521781e