From: Atari Ace [mailto:atari_ace@verizon.net] Sent: Monday, August 25, 2008 8:16 AM Subject: [patch] Remove useless backslashes
Hi mamedev,

I've been using pmd 4.2.2 to analyze the MAME codebase for duplicate
code, and it's C parser complains about useless backslashes and
dangling defines in #if...#endif sections.  Here's a patch that fixes
this, and additionally fixes one real bug, a string in djmain.c was
malformed, it's curious it compiled.

~aa



git-svn-id: svn://dspnet.fr/mame/trunk@2693 749742ba-7341-0410-aadc-df50b521781e