From: Samuele Zannoli [mailto:samuele.zannoli@airmachine.it] Sent: Tue 9/23/2008 1:16 PM To: submit@mamedev.org Subject: More naomi ... Hi.
This patch shows the contents of the framebuffer if the 3d accelerator is
not used so that now the atomiswave logo and messages are shown (if you
wait enough).
Removes a maple bug (no more strange messages in cvs2gd).
Improves documentation of the communication registers with the dimm board.
And if DEBUG_VERTICES is defined as 1 the vertices sent to the tile
accelerator are collected and then drawn connected by a segment to give a
wireframe-like view of the scene.

Bye,
 Samuele Zannoli



git-svn-id: svn://dspnet.fr/mame/trunk@2922 749742ba-7341-0410-aadc-df50b521781e