From: Phill Harvey-Smith [mailto:afra@aurigae.demon.co.uk] Sent: Saturday, October 11, 2008 12:01 PM To: submit@mamedev.org Subject: 6821 PIA changes
Hi,

Can this change be added to the 6821 pia code, it adds a function that 
allows the client code to get the status of CA2, taking account of the 
internal pull up resistor. This code only adds a function and does not 
change any existing code so should not have any impact on existing
drivers.

This change is needed for a couple of the mess drivers.

Thanks.

Phill.

-- 
Phill Harvey-Smith, Programmer, Hardware hacker, and general eccentric !



git-svn-id: svn://dspnet.fr/mame/trunk@3081 749742ba-7341-0410-aadc-df50b521781e