From: tbellm@arcor.de [mailto:tbellm@arcor.de] Sent: Sunday, October 19, 2008 9:01 AM To: submit@mamedev.org Subject: Sound Support for Cosmic Alien
Hello,

I own a cosmic alien PCB and so I thought it would be a good idea to add
sound support
to this game by changing the corresponding driver ("drivers\cosmic.c").
I also have recorded the different sound effects as wave - files.

I have attached two files to this mail. The file "cosmica.zip" contains
the samples
(so it can be simply copied to the "samples" folder) and
the other one "cosmica_diff.zip" contain the "diff" file taken against
the 0128 sources.
So you can check it and maybe you want to implement the changes in a
further release?

best regards
T.Bellm



git-svn-id: svn://dspnet.fr/mame/trunk@3145 749742ba-7341-0410-aadc-df50b521781e