From: Micko [mailto:mmicko@gmail.com] Sent: Saturday, December 13, 2008 2:15 AM To: Aaron Giles Subject: Re: Another 8080/85 change from me
Hi Arron,
 
Here is the patch.
 
Problem was in EI handling and not clearing of interrupt enable flag, as I did in previous. So now EI is done as in Z80 since it is working same way. I have tried a invaders clone having problem, now they work, also MESS machine for which I initialy did this works.
 
Hope this will be in u6.
 
Micko



git-svn-id: svn://dspnet.fr/mame/trunk@3815 749742ba-7341-0410-aadc-df50b521781e