From: Duke [mailto:startaq@gmail.com] Sent: Wednesday, February 25, 2009 5:20 AM To: submit@mamedev.org Subject: CIA update
Hello,

this patch adds support for the /PC output pin to the CIA emulation.
From the datasheet "/PC will go low for one cycle following a read or write of PORT B.". 
This is needed for centronics printer emulation on the MESS side, but I imagine it would 
simplify/correct the emulation in some MAME amiga drivers as well (mquake.c for example).

--Dirk



git-svn-id: svn://dspnet.fr/mame/trunk@4618 749742ba-7341-0410-aadc-df50b521781e