Moved chd_clone_metadata code into chdman.c so that it can intelligently update old tags.
Modified cdrom.c to expose metadata read/write functions, and changed
chdman to use them. Also changed chdman to parse old-style metadata and
convert it to new-style metadata since we're going to need to re-do this
anyway and the two won't hash to the same value.

Added completely untested support for the ident metadata to the hard
disk creation path. If a filename is provided immediately after the
source filename, it is taken as an ident file and processed for CHS
data and added. Other metadata types can be added afterwards via
-addmeta.


git-svn-id: svn://dspnet.fr/mame/trunk@4769 749742ba-7341-0410-aadc-df50b521781e