Having decided to look at the MAME source again, I have a fairly minor patch:
* Fix build of ldplayer on OS X.  Since the CUSTOM sound module no longer exists, I 
arbitrarily changed it to WAVE, as ar gets upset if it has no input files.  I also 
removed the -all_load flag for ldplayer from the main makefile as it upsets the linker 
on OS X.

* Fix build for PPC64 Linux.  (This slightly messes up static branch prediction hints 
on OS X and AIX, but OS X for PPC64 is dead, and no- one builds MAME for AIX, and 
it will still build, anyway.)

* Paramaterise the arguments to check for NULL in the ATTR_NONNULL macro rather than 
just checking the first argument.  This requires compiler support for C99 variadic 
macros (MSVC2005 and GCC4 have this AFAIK).

Vas




git-svn-id: svn://dspnet.fr/mame/trunk@4889 749742ba-7341-0410-aadc-df50b521781e