after further analysis this is what XingXing's changes actually do.
Added emulation of the earlier type of ARM (ASIC27A, 5585E type)
Used this for protection emulation on Photo Y2k, Knights of Valour Superheroes instead of old simulation code
This means KOV Superheroes now works.

Disabled protection simulation for several sets.
This demotes Knights of Valour / Knights of Valour Plus to NOT working as there is no dump of the internal rom.

Photo Y2k still works because it's using the emulation added above, but You can no longer change region as this is supplied by the ASIC27A device which is now emulated.  This is correct.

added several new sets, several of which are re-releases of the games using the ASIC 27A instead of the previous protection devices. (oldss, killbldp)

Removed per-game speedup hacks, as they're no longer needed for the ARM based games to boot correctly.




git-svn-id: svn://dspnet.fr/mame/trunk@4976 749742ba-7341-0410-aadc-df50b521781e