Simulate 'keyoff' in the SPU.
The envelopes aren't actually emulated, but the sound is silenced after a key-off (by placing the sound in a 'release' state)  This prevents audiable sound from looping forever.

Not extensively tested, but good for my sanity when running some of the gnet additions, and the tgmj siren looping.

I may consider doing a better implementation if I can manage to get my head around all the envelope stuff.



git-svn-id: svn://dspnet.fr/mame/trunk@5046 749742ba-7341-0410-aadc-df50b521781e