Hi again Aaron,
Remember me, I sent you the Aliens ASIAN EPROM dump a few days ago?

Anyhow, I've just been repairing an original Data 
East Robocop boardset and noticed that only the 
PROM at 2A on the ROM board had been dumped - the 
two PROMs on the CPU board were undumped. So I 
desoldered those two from my boardset and dumped 
them for you. They didn't have labels on them so 
the attached filename dumps are made up of the 
type of PROM followed by the IC location on the PCB.

I also noticed that the filename of the already 
dumped PROM at 2A (which I verified okay against 
mine) didn't include the PROM type, although it's mentioned here:

http://maws.mameworld.info/maws/driverinfo/dec0.c

so I renamed that file and re-sent it. Up to you 
or the MAME team to decide whether to change the 
name, although I guess it's no big deal, although 
it makes life a bit easier for anyone repairing a 
boardset and wishing to know the PROM type (the 
boardset itself is screened wrong I believe - for 
the PROM at 2A it's screened as MB7130, yet the 
PROM used (and mentioned on the above page) is in fact an MB7124).

For further info, this is what happens when the 
relevant PROMs are removed from the board (one at a time!):

MB7122E (EN-24) PROM at 17E on CPU board -- 
Removing this causes all text to disappear along 
with the sprites and the foreground layer of the 
background graphics to be absent (but the background layer is still visible)

MB7116E at 12C on CPU board -- Removing this causes all sprites to disappear

MB7124E at 12A on ROM Board (although the board 
is screened as MB7130) -- Removing this causes the 
display to disappear, although coining up makes 
the correct text appear (and sound to play), 
intro graphics appear on starting a game, then 
nothing, it won't even play blind (although the 
coin up sound still plays on coin insert)


There's also a PAL on the ROM board but 
unfortunately it's protection fuse is of course 
blown so I can't read it. Ah well.

If you need to credit me when these are 
incorporated into the code, just put in my name Irongiant

Hope all this helps.


Cheers,
Phil



git-svn-id: svn://dspnet.fr/mame/trunk@5545 749742ba-7341-0410-aadc-df50b521781e