From: David Haywood [neohaze@nildram.co.uk] Sent: Thursday, July 09, 2009 12:20 PM To: Aaron Giles Subject: Jclub2 update
added different machine drivers for the different sets.
added an st0016 to the 'older hardware' machine driver, that's what the
'z80' rom on it is for (see st0016.c)  the newer hardware and bootleg
don't use it.

Gerald said he'll provide more information on this, so hopefully the
real graphic roms can be dumped.



git-svn-id: svn://dspnet.fr/mame/trunk@6050 749742ba-7341-0410-aadc-df50b521781e