DIPSW fix [kanikani]
argus: fixed service switch (was defined to port that does not exist), and fixed DIPLOCATION order.

valtric: fixed note for how to enter test mode.
coin1 and coin2 were swapped, and fixed DIPLOCATION order.

butasan: coin1 and coin2 were swapped, and fixed DIPLOCATION order.

To enter test mode:
argus:   power ON while coin2 is ON
valtric: power ON while coin1 is ON
butasan: power ON while service SW is ON

I don't know how to map multiple MAME keys to 1 real port on PCB, so I mapped service switch of butasan to "TEST(F2)".
As a result of this, you have to press F2 twice to use as normal service switch.

git-svn-id: svn://dspnet.fr/mame/trunk@6167 749742ba-7341-0410-aadc-df50b521781e