39in1 updates: [R. Belmont, MooglyGuy]
- Video, inputs, and flash data ROM hooked up
- Improved PXA255 LCDC, DMA and I2S emulation
- Working but not great CPLD protection simulation


It now boots but fails after the initial POST (notes on that are at the
top of the driver - MG and I are chasing it, but more eyes is good too).

You can get into service mode with F2 and step through the service screens
with F2.  All of them behave except the input test, which will time out after
a little bit and say "System Check Error".

Delete your .nv if it doesn't start up.



git-svn-id: svn://dspnet.fr/mame/trunk@6302 749742ba-7341-0410-aadc-df50b521781e