From: hoge hoge Date: Sat, 25 Jul 2009 11:57:29 -0700 To: submit@mamedev.org<submit@mamedev.org> Subject: 68k cpu reset eat cycles fix Hello,
Attached is a diff for 0133 that fixes a bug related to eating cycles 
during 68k cpu reset. Previously, initial_cycles had a wrong value 
after reset, and if cycles were < 0, it'd always return 0, .. check 
the diff to see what I mean.

This change also happens to fix a scrolling bug in Fantasy Zone, 
introduced when that cycle eating thing was added.

Greets,
hap




git-svn-id: svn://dspnet.fr/mame/trunk@6304 749742ba-7341-0410-aadc-df50b521781e