From: David Haywood [neohaze@nildram.co.uk] Sent: Friday, August 07, 2009 8:49 AM To: Aaron Giles Subject: Qix alt set
Klaus / Team Europe dumped this alt Qix set.

Qix (set 2, smaller roms) [Team Europe]

It actually appears to be the same set as one of the supported ones, but
with smaller roms.  This one is definitely from an original Taito board
(he sent a picture), I'm not so sure about the other set because it's
the odd one out as far as rom sizes go.  I've left both in the driver
for best reference.

I've also renamed 'qixa' to 'qixo' as it seems to be an older version
and tagged it as such.  Language selection is disabled, and there is a
spelling error on the screen not present in the other sets.  Please
update mamdriv.c accordingly.



git-svn-id: svn://dspnet.fr/mame/trunk@6455 749742ba-7341-0410-aadc-df50b521781e