From: Sandro Ronco [mailto:sandro.ronco@alice.it] Sent: Wednesday, August 26, 2009 5:57 AM To: submit@mamedev.org Subject: MAME cheat search engine
This is a diff of my cheat search engine with help of Pugsy.  
This is only a first part, not has the same functions of the old search engine, but is better than nothing

I have update the search engine to support search of byte, word, dword
and qword signed and unsigned.



git-svn-id: svn://dspnet.fr/mame/trunk@6673 749742ba-7341-0410-aadc-df50b521781e