Added new function mame_fclose_and_open_next() which will close a file and then keep searching the searchpath for the next valid file. Did some internal rearranging in fileio.c to make this work.
Changed cheat search so that it loads *all* cheat files in all search
paths. Note that it is easy to end up with duplicate entries this way.
Some currently disabled code is present which filters out duplicates,
but the logic for doing this is quite unclear with the presence of
text-only cheats, which is why the code is disabled for now.


git-svn-id: svn://dspnet.fr/mame/trunk@6677 749742ba-7341-0410-aadc-df50b521781e