From: David Haywood [neohaze@nildram.co.uk] Sent: Friday, September 25, 2009 12:15 PM To: Aaron Giles Subject: 360 pad crash
This 'fixes' the bug where MAME crashes and corrupts your config files
if you connect a pad at runtime.  Previously there was an assert, now it
simply doesn't execute that block of code if devcode is null.

I don't know if this is the best fix (it would probably be quite nice if
it could detect pads being added removed at runtime and adapt to that)
but at least it stops it crashing and corrupting files.



git-svn-id: svn://dspnet.fr/mame/trunk@6927 749742ba-7341-0410-aadc-df50b521781e