From: David Haywood [neohaze@nildram.co.uk] Sent: Wednesday, September 30, 2009 2:43 PM To: Aaron Giles Cc: Klaus Sommer Subject: Re: Unknown dump (galaxian hardware)
I'll leave this one up to Aaron to decide if it's worth including
anyway.  I've attached the diff for you.  Note, the current emulation of
this pacmanbl set doesn't seem to work properly at the moment anyway
(since the galaxian rewrite), the sprites are offset and vanish at the
top of the screen.

David Haywood wrote:
> It's a pacman bootleg, the content is identical to the pacmanbl set in
> MAME, although the order of the data in the roms is different.  I'm
> not sure it will be supported.
>
> Klaus Sommer wrote:
>> hi!
>>
>> attached are the dumps of an unknown board (looks like galaxian
>> hardware)...i think an bprom is missing, but i'm not sure which one
>> it is!
>>



git-svn-id: svn://dspnet.fr/mame/trunk@6957 749742ba-7341-0410-aadc-df50b521781e