A bottle of champaign to the person who explains the speed improvement in dkong from 1230% to 1470% due to this change (setting context->data to 0 in reset). Actually, it will avoid the first stream_update of buffer_stream when after approx. 11 samples the dac is written. Without setting data to 0, stream_update(buffer_stream) is called before stream_update(discrete_stream) is called. With the patch, the order is of the stream_updates is determined by streams.c.
What makes this interesting is, that the startup order of the stream_updates
determines the whole run-time of the game. When profiling, NODE_73 (Sallen-Key-Filter)
needs 1066 cycles without the patch and under 135 with the patch, although it is the 4th node
after DISCRETE_INPUT_BUFFER.

All in all this looks like a caching issue, however I am currently clueless what in the end
is causing it.


git-svn-id: svn://dspnet.fr/mame/trunk@7016 749742ba-7341-0410-aadc-df50b521781e