New game added - Not working: Super Hang-on Japan ver 317-0038
Don't credit me in the whatsnew

This is only the rom defs, the former "bootleg upgrade" looks to be a bootleg of this new official set.

Some who knows what they're doing needs to hook up the fd1094 decryption routine and check the CPU1 roms are loading in correctly.  The CPU2 roms should be correct.

credit those listed by Yasu in his original email describing the upload please.

git-svn-id: svn://dspnet.fr/mame/trunk@7067 749742ba-7341-0410-aadc-df50b521781e