more rom name corrections for the merit.c driver [Brian Troha]
don't mention this part:
I had to download the zips and info text files from softwarethis.com and compare the CRC32 values of the file then match the file with info for the correct rom label. Big thumbs down for the naming convention when dumping the roms (IE: metit2_1.1 ect), but at least a thumbs up for making and keeping the info files with correct label info.

git-svn-id: svn://dspnet.fr/mame/trunk@7685 749742ba-7341-0410-aadc-df50b521781e