Removed deprecated core function decodegfx() [Christophe Jaillet]
---------- Forwarded message ----------
From: Christophe Jaillet <christophe.jaillet@wanadoo.fr>
Date: Mon, Dec 14, 2009 at 9:56 PM
Subject: Removal of a deprecated function
To: submit@mamedev.org


Hi,

a function in "drawgfx.c" is declared in "deprecat.h"
This function ("decodegfx(...)") is not used any more in the code, so it
could be removed.

Hope this help.
Best regards,

CJ

git-svn-id: svn://dspnet.fr/mame/trunk@7692 749742ba-7341-0410-aadc-df50b521781e