hng64: bbust2 now has 3d geometry. [Andrew Gardner]
NOTES:
The geometry is correct in bbust2 (and it's 'playable' - I beat the first stage), 
but it's missing half of its geo and the textures are wrong.  They're close to 
correct though.  I think there are multiple texture page sizes that I'm not taking into account.

I just noticed there might be a crash or two still lurking in there as well.  Will fix soon.

git-svn-id: svn://dspnet.fr/mame/trunk@7865 749742ba-7341-0410-aadc-df50b521781e