Added a PCM sample end/loop callback to the RF5C68 core [Barry Rodewald]
---------- Forwarded message ----------
From: Barry Rodewald <bsr@xnet.co.nz>
Date: Sat, Jan 2, 2010 at 10:13 AM
Subject: RF5C68 addition
To: submit@mamedev.org


Hi,

Here is a small update for the RF5C68 PCM core.  For the FM Towns MESS
driver, there needs to be some way to tell when a PCM sample has ended
or looped, as the system can generate an interrupt at this point,
usually to end the sample playback on the specified channel.  From the
RF5C68(A) datasheet, there appears to be no obvious way for an IRQ
signal to be sent, so I figure the PCM IRQ setup must be external to
the PCM chip itself.

Thus, I've added a simple callback, which should be called when the
end of a sample is reached, and with a parameter for the device
itself, and the channel on which the sample has ended.

Example of the result, with callback implemented:
http://mahlemiut.marpirc.net/mess_raiden.mp3  (sound effects are done
by the RF5C68, music is CD-DA).

Thanks,
Barry Rodewald
mailto:bsr@xnet.co.nz

git-svn-id: svn://dspnet.fr/mame/trunk@7970 749742ba-7341-0410-aadc-df50b521781e