Add support for per-scanline scroll/zoom and per-column effects. This is adds a significant performance penalty when enabled. Removed warning flag from s1945ii and s1945iii. s1945iii needs confirmation of correctness and x-zoom is a guess. [PsikyoFan]
This can be seen on:
* s1945ii level 7 intro, level 8 boss. (Hold PL1 Button 1 which hitting F2 for stage select).
* s1945iii level 7 intro, level 7 main, level 8 intro, level 8 boss (Enter "8-1-6-5-0" as code in test mode for stage select).

Note that this is slow. Suggestions for speeding it up are welcome.
Rendering each tilemap to an ARGB bitmap, including applying the per-pen or fixed alpha.
Extracting an entire scanline from the bitmap, transforming it and then blending the ARGB scanline.
Obvious optimisations for transparency and fixed alpha in place.

Alternative is to just try and draw with drawgfx directly and fiddle with cliprects, but zooming makes that impossible to get nice without fudging the scaling/positioning slightly and unlikely to be faster.

Hack for daraku is still there, working to get rid of it. Aside from above and MT03663 games are now believed to be pixel-perfect.


git-svn-id: svn://dspnet.fr/mame/trunk@8036 749742ba-7341-0410-aadc-df50b521781e