Make mips register names accessible in the debugger expressions [Olivier Galibert]
The trick is to lookup the register name before treating '$' as
announcing an hex value.  It doesn't make anything impossible or even
hard since even if you have the register name '$a0' you can still use
a bare 'a0' or an explicit '0xa0' to get to the value.  Symmetrically,
'$a0' still works for the hex value on say m68k with 'a0' as the
register.

git-svn-id: svn://dspnet.fr/mame/trunk@8369 749742ba-7341-0410-aadc-df50b521781e