Fixed HD63701 SLP opcode [Sandro Ronco]
---------- Forwarded message ----------
From: Sandro Ronco <sandro.ronco@gmx.com>
Date: Mon, Jul 26, 2010 at 6:07 PM
Subject: [HD63701]Fix slp opcode
To: submit@mamedev.org


Fixed SLP opcode basing on HD63701 and HD6303 datasheet.
This is need for psion driver in MESS, but I did some testing in MAME
and I have not seen regressions.
Probably nobody noticed this before because this opcode is almost
unused in arcade machine.

If you want have a confirmation of this you can see the diagram at
page 14 of the HD63701 datasheet or at page 24 of the HD6303
datasheet, where it is clear that the "sleep cancel signal" is not
conditional at the IRQ mask.

Regards

Sandro Ronco

git-svn-id: svn://dspnet.fr/mame/trunk@9828 749742ba-7341-0410-aadc-df50b521781e