New Working set: Raiden Fighters (Asia?, Metrotainment license) [anondumper]
Marked many roms on the two asia? sets of raiden fighters as bad pending redump, as the rom layout on the asian? spi pcb is different and several roms were clearly recombined into single roms. These need to be redumped for verification that the contents are the same.

renamed rdftdi to rdftadi

git-svn-id: svn://dspnet.fr/mame/trunk@9917 749742ba-7341-0410-aadc-df50b521781e