New build tool makedep.c, derived from src2html.c (really!). Does basic dependency analysis and supports excluding paths.
Makefile now supports a depend target which will run makedep
on the src/$(TARGET) tree and create dependencies. It deliberately
excludes the root of src/emu as well as the osd directories in
order to keep the dependency file down to a reasonable size, so
if a core header file changes, you still need to clean and build
all.


git-svn-id: svn://dspnet.fr/mame/trunk@10273 749742ba-7341-0410-aadc-df50b521781e