Put the SH2 DMA on a timer.  This is needed because Knuckles Chaotix on the 32X can't be done with 'Instant DMA' due the SH2 DMA is used to read from a FIFO port which is being filled by the 68k during the transfer.  [David Haywood]
Haze: "I need to review save state support in the core again, a few extra things probably need saving now.

It might be possible to avoid the high frequency timer if it causes too many performance issues by moving the update of the DMA to the EXECUTE loop instead.  
(I want to experiment with this before updating the save state support, however if somebody else wants to fix savestates in the meantime they're welcome, 
as long as they don't get offended if their code gets changed...)"


git-svn-id: svn://dspnet.fr/mame/trunk@10339 749742ba-7341-0410-aadc-df50b521781e