tms57002: significantly improved build speed on lower-end h/w [R. Belmont]
No-whatsnew explanation: this takes OPTIMIZE=3 compile time for this core from 
over 1 hour down to about 2 minutes on PS3 Linux.  (Yes, the tms57002 itself 
took over an hour to compile before - the thing has 256 MB of RAM and 
a very slow HDD so when it hits swap, swap hits back).



git-svn-id: svn://dspnet.fr/mame/trunk@11475 749742ba-7341-0410-aadc-df50b521781e