New Clones -------------- Ropeman (bootleg of Roc'n Rope)
Comment:
Yes, this is that one.
It's been sourced numerous times over the years and it's a very popular bootleg, even if the changes are minimal.  It was time to add back to the documentation.  Cool thing is that it is fully playable now when it did not when it was taken out of MAME in 0.36b10. :)

git-svn-id: svn://dspnet.fr/mame/trunk@11896 749742ba-7341-0410-aadc-df50b521781e