The art of destruction. (nw)
Part II: Artwork.
aperture.png: Based on Aaron's aperture_4x6 mask but with more prominent notching to bring out the "brick" pattern that was used on many period colour monitors, including the Commodore 1702 - which was, admittedly, not an arcade monitor, but it looks believable and can be changed.
white.png: Default 8x8 255,255,255,255 texture.

git-svn-id: svn://dspnet.fr/mame/trunk@12545 749742ba-7341-0410-aadc-df50b521781e