memory: Implement depopulate_unused().  [O. Galibert]
depopulate_unused is essentially the garbage collector for
bottom-level handlers.  The previous code was working around not
having it implemented yet by reusing handlers with the same start
address, end address and address mask.  The problem with that trick is
that it is slightly incorrect.  If you have a memory map with:
  AM_RANGE(0x0000, 0x0fff) AM_READ(up_r)
  AM_RANGE(0x0000, 0x0fff) AM_READ(down_r) AM_MIRROR(0x8000)

then the range 8000..8fff would have called up_r instead of down_r due
to the handler reuse.  The mirror value is not saved, hence not
compared.  New code doesn't use the trick, so doesn't call the wrong
handler, but OTOH eats handlers for breakfast.

It's important to note that such a memory map is so highly improbable
hardware-wise that it wasn't worth worrying about.  But it's a
behaviour change it is interesting to keep in mind.  In particular
since the up_r range can be added through an install_read_handler
instead of a static map.

git-svn-id: svn://dspnet.fr/mame/trunk@12773 749742ba-7341-0410-aadc-df50b521781e