MESS version of MAME r14882 (includes the MAME changes files as well):
Bulk converted cliprect * to cliprect & across the system. This makes
cliprects mandatory everywhere. In general, cliprects were being
correctly passed through the video side of most drivers already, so
it is mostly a semantic change. Note that with my previous change,
bitmaps have cliprects, so if you just want to clip to the bitmap's
boundaries, pass bitmap->cliprect() instead of NULL (which is no 
longer permitted). [Aaron Giles]

Note that in MESS there were a number of drivers ignoring the 
cliprect passed to SCREEN_UPDATE. This is now rectified.



git-svn-id: svn://messdev.no-ip.org/mess@13799 9849079f-6229-0410-a07f-84406ce305ba