MESS equivalent of MAME 15022:
Some cleanup of the previous change. Added new template
device_delegate which wraps a regular delegate and includes
a string pointer to a device tag, which can be simply
resolved later. Converted the screen_update delegates to
to be based on this. Changed the mechanism by which screen
formats are auto-deduced. Converted SCREEN_EOF to use these
delegates as well, so now there is MCFG_SCREEN_EOF_STATIC/
DRIVER/DEVICE just like MCFG_SCREEN_UPDATE.



git-svn-id: svn://messdev.no-ip.org/mess@14016 9849079f-6229-0410-a07f-84406ce305ba