Partial rewrite of raiden2's sprite decryption. No whatsnew.
This is functionally equivalent to the previous code, but it show more clearly the structure of the nonlinear part of the current algorithm, showing the SPI-like sum with conditional carry and a second bunch of "carries" (another SPI-like sum?). This should provide a better base for research that the previous implementation, and it points more clearly to how the real algorithm works.

git-svn-id: svn://dspnet.fr/mame/trunk@15727 749742ba-7341-0410-aadc-df50b521781e