New game added and new Blood Storm Clone
Pairs Red is basically like the Pairs game but is designed for children and is a ticket redemption game so I consider it different enough to be a new game. Although the matching of pairs is the basic concept for both.

New Clone Added
-------------------------------
Blood Storm v1.04 [Brian Troha, The Dumping Union]

New Game Added
--------------------------
Pairs Red (V1.0, 10/25/94) [Brian Troha, The Dumping Union]


git-svn-id: svn://dspnet.fr/mame/trunk@17094 749742ba-7341-0410-aadc-df50b521781e