Apple2C added. Doesn't work in the browser, though.
Fixed the buildtools target in the JSMESS Makefile. I need to go back and make the Makefile a bit more sane.

Caveat: Apple2C does not build properly since the 'buildtools' target for MESS does not build a buildtool that is only used to build the CPU core for the M68K. I looked and found a similar tool for a different CPU. Looks like I'll have to fix the MESS buildtools target, or have the JSMESS buildtools target compile a more "complete" MESS.